Move ContainsGenesis to separate file
Avoids some module interdependencies.
